Fix schema
Schema 0 had a column named "user", which is a reserved word in some
databases (eg, postgres). This column has been later renamed to
"user_id" with a new plugin version (steps 1 to 9). However, since the
installation for a new instance starts from step 0, the process fails.

If this is a new installation, we can skip the original table and
subsequent upgrade and create directly the new table structure.